Improve ELD text
There were some typos and mistakes with the links in the table of contents.
The typos meant that the card modal did not work correctly with the cards.
One thing in particular that is a bit annoying is the fact that adventures
need both halves spelt out. This may be improved.